backendeggmobile/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
boto3==1.35.21
botocore==1.35.21
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-storages==1.14.4
djangorestframework==3.15.2
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.8.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
